separate files for multimodal and text impelmentation, separate utility to load_video
