+++ samples/cpp/visual_language_chat/video_to_text_chat.cpp
@@ -2,87 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <openvino/genai/visual_language/pipeline.hpp>
-#include <openvino/genai/visual_language/video_metadata.hpp>
-#include <opencv2/core.hpp>
-#include <opencv2/videoio.hpp>
+#include "load_video.hpp"
 #include <iostream>
-#include <filesystem>
-#include <set>
-#include <cstring>
-
-namespace fs = std::filesystem;
-
-std::vector<size_t> make_indices(size_t total_frames, size_t num_frames) {
-    std::vector<size_t> indices;
-    indices.reserve(num_frames);
-
-    auto step = float(total_frames) / num_frames;
-
-    for (size_t i = 0; i < num_frames; ++i) {
-        size_t idx = std::min(size_t(i * step), total_frames - 1);
-        indices.push_back(idx);
-    }
-
-    return indices;
-}
-
-std::pair<ov::Tensor, ov::genai::VideoMetadata> load_video(const std::filesystem::path& video_path, size_t num_frames = 8) {
-    cv::VideoCapture cap(video_path.string());
-
-    if (!cap.isOpened()) {
-        OPENVINO_THROW("Could not open the video file.");
-    }
-
-    const size_t total_num_frames = cap.get(cv::CAP_PROP_FRAME_COUNT);
-    const auto frames_indices = make_indices(total_num_frames, num_frames);
-
-    ov::genai::VideoMetadata video_metadata;
-    video_metadata.fps = cap.get(cv::CAP_PROP_FPS);
-    // Passing video metadata with frame indices defined enables sampling based on provided indices within the pipeline,
-    // and any model-specific sampling logic will be skipped (if defined).
-    // Leave frames_indices empty to apply model-specific sampling (e.g. for Qwen3-VL).
-    video_metadata.frames_indices = frames_indices;
-
-    std::vector<cv::Mat> frames;
-    cv::Mat frame;
-    size_t width  = cap.get(cv::CAP_PROP_FRAME_WIDTH);
-    size_t height = cap.get(cv::CAP_PROP_FRAME_HEIGHT);
-    ov::Tensor video_tensor(ov::element::u8, ov::Shape{total_num_frames, height, width, 3});
-    auto video_tensor_data = video_tensor.data<uint8_t>();
-
-    size_t frame_idx = 0;
-    while (cap.read(frame)) {
-        OPENVINO_ASSERT(frame.cols == width && frame.rows == height && frame.channels() == 3);
-        memcpy(video_tensor_data, frame.data, frame.total() * 3 * sizeof(uint8_t));
-        video_tensor_data += frame.total() * 3;
-        frame_idx++;
-    }
-    OPENVINO_ASSERT(frame_idx == total_num_frames,
-        "Frame count mismatch: expected " + std::to_string(total_num_frames) + ", got " + std::to_string(frame_idx));
-    
-    return {video_tensor, video_metadata};
-}
-
-std::pair<std::vector<ov::Tensor>, std::vector<ov::genai::VideoMetadata>> load_videos(
-    const std::filesystem::path& input_path
-) {
-    if (input_path.empty() || !fs::exists(input_path)) {
-        OPENVINO_THROW("Path to videos is empty or does not exist.");
-    }
-    if (fs::is_directory(input_path)) {
-        std::set<fs::path> sorted_videos{fs::directory_iterator(input_path), fs::directory_iterator()};
-        std::vector<ov::Tensor> videos;
-        std::vector<ov::genai::VideoMetadata> videos_metadata;
-        for (const fs::path& dir_entry : sorted_videos) {
-            const auto [video, video_metadata] = load_video(dir_entry);
-            videos.push_back(video);
-            videos_metadata.push_back(video_metadata);
-        }
-        return {videos, videos_metadata};
-    }
-    const auto [video, video_metadata] = load_video(input_path);
-    return {{video}, {video_metadata}};
-}
 
 ov::genai::StreamingStatus print_subword(std::string&& subword) {
     std::cout << subword << std::flush;
@@ -93,7 +14,7 @@ int main(int argc, char* argv[]) try {
     if (argc < 3 || argc > 4) {
         OPENVINO_THROW(std::string{"Usage "} + argv[0] + " <MODEL_DIR> <VIDEO_FILE OR DIR_WITH_VIDEOS> <DEVICE>");
     }
-    const auto [videos, videos_metadata] = load_videos(argv[2]);
+    const auto [videos, videos_metadata] = utils::load_videos(argv[2]);
 
     // GPU and NPU can be used as well.
     // Note: If NPU is selected, only language model will be run on NPU
